Move upload backend into library
